requirements.txt
streamlit==1.41.1
pandas
mysql-connector-python
python-dotenv
groq
mdpdf
pymupdf
